+++ src/screens/home/Home.js
@@ -4,6 +4,7 @@ import RestaurantCard from './RestaurantCard'
 import axios from 'axios'
 import { useHistory } from 'react-router-dom'
 import { SearchField } from './SearchField'
+import { HeaderTop } from '../../components/HeaderTop/HeaderTop'
 
 
 const baseUrl = 'https://us-central1-missao-newton.cloudfunctions.net/futureEatsA'
@@ -97,12 +98,7 @@ function Home() {
 
     return (
       <Main>
-        <Header>
-          <BackButton/>
-          <Title>
-            <TitleIn> FutureEats </TitleIn>
-          </Title>
-        </Header>
+        <HeaderTop backButton={false} title={'FutureEats'} />
         <PageBox>
           <SearchField history={history} />
           <FilterBar> {renderCategories()} </FilterBar>
